add fullscreen recording (nice)

+++ src/screen_recorder.py
@@ -57,7 +57,11 @@ def start_record(metadata: MetadataObject, output: str):
         if metadata.cursor:
             process = (
                 ffmpeg
-                .input(input_device, format=input_format, framerate=metadata.fps)
+                .input(
+                    input_device,
+                    format=input_format,
+                    framerate=metadata.fps
+                       )
                 .output(f"{output}/{metadata.title}.mp4", vcodec="libx264", pix_fmt="yuv420p")
                 .overwrite_output()
                 .run_async(pipe_stdin=True)
@@ -65,7 +69,12 @@ def start_record(metadata: MetadataObject, output: str):
         else:
             process = (
                 ffmpeg
-                .input(input_device, format=input_format, framerate=metadata.fps, draw_mouse=0)
+                .input(
+                    input_device,
+                    format=input_format,
+                    framerate=metadata.fps,
+                    draw_mouse=0
+                )
                 .output(f"{output}/{metadata.title}.mp4", vcodec="libx264", pix_fmt="yuv420p")
                 .overwrite_output()
                 .run_async(pipe_stdin=True)
@@ -74,11 +83,31 @@ def start_record(metadata: MetadataObject, output: str):
 
 def stop_record(process):
     """
-    Остановка фонового процесса записи (хотя подходит для любых процессов)
+    Остановка фонового процесса записи
     """
-    process.terminate() # спокойная остановка процесса
 
-    try:
-        process.wait(timeout=2) # ожидание остановки
-    except subprocess.TimeoutExpired:
-        process.kill() # если процесс не остановился, force stop
+    if process.poll() is None:  # если процесс жив
+        try: # по умолчанию ffmpeg останавливает запись через q
+            process.stdin.write(b'q\n')
+            process.stdin.flush()
+            process.wait(timeout=2)
+            return
+        except (BrokenPipeError, subprocess.TimeoutExpired):
+            pass
+
+        try: # если не помогло, пытаемся юзать ctrl+C
+            process.send_signal(0x40010003)  # CTRL_C_EVENT
+            process.wait(timeout=2)
+            return
+        except:
+            pass
+
+        try: # пытаемся остановить процесс программно
+            process.terminate()
+            process.wait(timeout=1)
+            return
+        except subprocess.TimeoutExpired:
+            pass
+
+        process.kill() # просто убиваем его, если ничего выше не помогло
+        process.wait()
